pyproject.toml
[project]
name = "agnolotti"
version = "0.1.0"
description = "Agentic AI project built with Agno and Claude Sonnet, served via AgentOS"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "agno[anthropic,os]>=2.5.0",
    "fastapi[standard]>=0.115.0",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.1.0",
    "sqlalchemy>=2.0.0",
    "pgvector>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv.pip]
no-annotate = true
